+++ src/Farmer/Common.fs
@@ -952,6 +952,33 @@ module Identity =
         static member (+) (managedIdentity, userAssignedIdentity:UserAssignedIdentity) =
             { managedIdentity with UserAssigned = userAssignedIdentity :: managedIdentity.UserAssigned }
 
+module Containers =
+    type DockerImage =
+        | PrivateImage of RegistryDomain : string * ContainerName : string *  Version:string option
+        | PublicImage of ContainerName:string * Version:string option
+            member this.ImageTag =
+                match this with
+                | PrivateImage (registry, container, version) ->
+                    let version = version |> Option.defaultValue "latest"
+                    $"{registry}/{container}:{version}"
+                | PublicImage (container, version) ->
+                    let version = version |> Option.defaultValue "latest"
+                    $"{container}:{version}"
+        /// Parses an image tag into a DockerImage record.
+        static member Parse (tag:string) =
+            match tag.Split([|':'|], StringSplitOptions.RemoveEmptyEntries) with
+            | [| repo; version |] ->
+                match repo.Split([|'/'|], StringSplitOptions.RemoveEmptyEntries) |> List.ofArray with
+                | first::rest when (first.Contains ".") -> DockerImage.PrivateImage(first, (rest |> String.concat "/"), Version=Some version)
+                | _ -> DockerImage.PublicImage(repo, Version=Some version)
+                | _ -> raiseFarmer $"Malformed docker image tag - incorrect number of repository segments: '{tag}'"
+            | [| repo |] ->
+                match repo.Split([|'/'|], StringSplitOptions.RemoveEmptyEntries) |> List.ofArray with
+                | first::rest when (first.Contains ".") -> DockerImage.PrivateImage(first, (rest |> String.concat "/"), None)
+                | _ -> DockerImage.PublicImage(repo, None)
+                | _ -> raiseFarmer $"Malformed docker image tag - incorrect number of repository segments: '{tag}'"
+            | _ -> raiseFarmer $"Malformed docker image tag - incorrect number of version segments: '{tag}'"
+
 /// Credential for accessing an image registry.
 type ImageRegistryCredential =
     { Server : string
@@ -1429,6 +1456,10 @@ module CosmosDb =
     /// A request unit.
     [<Measure>]
     type RU
+    /// The throughput for CosmosDB account
+    type Throughput =
+        | Provisioned of int<RU>
+        | Serverless
 
 module PostgreSQL =
     type Sku =
@@ -2103,34 +2134,21 @@ module ContainerApp =
     type ServiceBusScaleRule = { QueueName : string; MessageCount: int; SecretRef : string }
     type HttpScaleRule = { ConcurrentRequests : int }
     type StorageQueueScaleRule = { QueueName : string; QueueLength: int; StorageConnectionSecretRef : string; AccountName : string }
-    type CpuScaleRule = CpuUtilisation of int | CpuAverageValue of int
-    type MemoryScaleRule = MemoryUtilisation of int | MemoryAverageValue of int
+    type UtilisationRule = { Utilisation : int }
+    type AverageValueRule = { AverageValue : int }
+    type MetricScaleRule = Utilisation of UtilisationRule | AverageValue of AverageValueRule
     [<RequireQualifiedAccess>]
     type ScaleRule =
         | EventHub of EventHubScaleRule
         | ServiceBus of ServiceBusScaleRule
         | Http of HttpScaleRule
-        | CPU of CpuScaleRule
-        | Memory of MemoryScaleRule
+        | CPU of MetricScaleRule
+        | Memory of MetricScaleRule
         | StorageQueue of StorageQueueScaleRule
         | Custom of obj
     type Transport = HTTP1 | HTTP2 | Auto
-    type Visibility = External | Internal
+    type IngressMode = External of port:uint16 * Transport option | InternalOnly
     type ActiveRevisionsMode = Single | Multiple
-    type DockerImageKind =
-        | PrivateImage of RegistryDomain : string * ContainerName : string *  Version:string option
-        | PublicImage of ContainerName:string * Version:string option 
-            member this.ImageTag =
-                match this with
-                | PrivateImage (registry, container, version) ->
-                    let version = version |> Option.defaultValue "latest"
-                    $"{registry}/{container}:{version}"
-                | PublicImage (container, version) ->
-                    let version = version |> Option.defaultValue "latest"
-                    $"{container}:{version}"
-
-    
-
 
 namespace Farmer.DiagnosticSettings