+++ src/Tests/ContainerApps.fs
@@ -10,7 +10,6 @@ let fullContainerAppDeployment =
     let containerLogs = logAnalytics { name "containerlogs" }
     let containerRegistryDomain = "myregistry.azurecr.io"
     let containerRegistryUsername = "myregistry"
-    let containerRepository = "myimage"
     let version = "1.0.0"
     let containerEnv =
         containerEnvironment {
@@ -19,44 +18,44 @@ let fullContainerAppDeployment =
             add_containers [
                 containerApp {
                     name "http"
-                    active_revision_mode ActiveRevisionsMode.Single
+                    active_revision_mode Single
                     add_registry_credentials [
                         registry containerRegistryDomain containerRegistryUsername
                     ]
                     add_containers [
                         container {
-                            container_name "http"
+                            name "http"
                             private_docker_image containerRegistryDomain "http" version
-                            cpu_cores 0.5<VCores>
+                            cpu_cores 0.25<VCores>
+                            memory 0.5<Gb>
                         }
                     ]
                     replicas 1 5
                     add_env_variable "ServiceBusQueueName" "wishrequests"
                     add_secret_parameter "servicebusconnectionkey"
-                    ingress_visibility External
+                    ingress_state Enabled
                     ingress_target_port 80us
                     ingress_transport Auto
                     dapr_app_id "http"
-                    add_scale_rule "http-rule" (ScaleRule.Http { ConcurrentRequests = 100 })
+                    add_http_scale_rule "http-rule" { ConcurrentRequests = 100 }
                 }
                 containerApp {
                     name "servicebus"
-                    active_revision_mode ActiveRevisionsMode.Single
+                    active_revision_mode Single
                     add_containers [
                         container {
-                            container_name "servicebus"
+                            name "servicebus"
                             private_docker_image containerRegistryDomain "servicebus" version
                         }
                     ]
                     replicas 0 3
                     add_env_variable "ServiceBusQueueName" "wishrequests"
                     add_secret_parameter "servicebusconnectionkey"
-                    add_scale_rule
-                        "sb-keda-scale" 
-                        (ScaleRule.ServiceBus {
-                            QueueName = "wishrequests"
-                            MessageCount = 5
-                            SecretRef = "servicebusconnectionkey" })
+                    add_servicebus_scale_rule
+                        "sb-keda-scale"
+                        { QueueName = "wishrequests"
+                          MessageCount = 5
+                          SecretRef = "servicebusconnectionkey" }
                 }
             ]
         }
@@ -71,7 +70,7 @@ let tests = testList "Container Apps" [
     let jobj = JObject.Parse jsonTemplate
 
     test "Container automatically creates a log analytics workspace" {
-        let env : IBuilder = containerEnvironment { name "testca" } :> _
+        let env : IBuilder = containerEnvironment { name "testca" }
         let resources = env.BuildResources Location.NorthEurope
         Expect.exists resources (fun r -> r.ResourceId.Name.Value = "testca-workspace") "No Log Analytics workspace was created."
     }
@@ -84,7 +83,7 @@ let tests = testList "Container Apps" [
     test "Full container environment kubeEnvironment" {
         let kubeEnv = jobj.SelectToken("resources[?(@.name=='kubecontainerenv')]")
         Expect.equal (kubeEnv.["type"] |> string) "Microsoft.Web/kubeEnvironments" "Incorrect type for kuberenetes environment"
-        Expect.equal (kubeEnv.["kind"] |> string) "containerenvironment" "Incorrect kind for kuberenetes environment"        
+        Expect.equal (kubeEnv.["kind"] |> string) "containerenvironment" "Incorrect kind for kuberenetes environment"
         let kubeEnvAppLogConfig = jobj.SelectToken("resources[?(@.name=='kubecontainerenv')].properties.appLogsConfiguration")
         Expect.equal (kubeEnvAppLogConfig.["destination"] |> string) "log-analytics" "Incorrect type for app log config"
         let kubeEnvLogAnalyticsCustomerId = jobj.SelectToken("resources[?(@.name=='kubecontainerenv')].properties.appLogsConfiguration.logAnalyticsConfiguration")
@@ -116,7 +115,7 @@ let tests = testList "Container Apps" [
         let httpContainer = containers |> Seq.head
         Expect.equal (httpContainer.["image"] |> string ) "myregistry.azurecr.io/http:1.0.0" "Incorrect container image"
         Expect.equal (httpContainer.["name"] |> string ) "http" "Incorrect container name"
-        Expect.equal (httpContainer.SelectToken("resources.cpu") |> float ) 0.5 "Incorrect container cpu resources"
+        Expect.equal (httpContainer.SelectToken("resources.cpu") |> float ) 0.25 "Incorrect container cpu resources"
         Expect.equal (httpContainer.SelectToken("resources.memory") |> string ) "0.50Gi" "Incorrect container memory resources"
 
         let scale = httpContainerApp.SelectToken("properties.template.scale")